backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
openai
python-dotenv
python-jose[cryptography]
passlib
